+++ manager/src/main/java/org/lsposed/lspatch/ui/component/SearchBar.kt
@@ -25,7 +25,9 @@ import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import org.lsposed.lspatch.TAG
+
+
+private const val TAG = "SearchBar"
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
